Merge upstream

+++ src/communication/incoming/messenger/MESSENGERINIT.h
@@ -19,7 +19,7 @@ void MESSENGERINIT(session *p, incoming_message *message) {
     om_write_int(friends_list, 600);
     om_write_int(friends_list, 200);
     om_write_int(friends_list, 600);
-    om_write_int(friends_list, list_size(p->messenger->friends)); // Buddy list count
+    om_write_int(friends_list, (int) list_size(p->messenger->friends)); // Buddy list count
 
     for (int i = 0; i < list_size(p->messenger->friends); i++) {
         messenger_entry *friend;
